begin refactor of plugin, remove ECS logic
