+++ servinator-plugin/ServinatorPlugin/src/main/resources/config.yml
@@ -1,0 +1,1 @@
+interval: 5
