--- servinator-plugin/app/src/main/resources/plugin.yml
@@ -1,7 +1,0 @@
-name: ServinatorPlugin
-version: 0.1.1
-main: com.github.sharp378.paper.servinator.ServinatorPlugin
-description: Handles server termination and cleanup.
-author: sharp378
-website: https://github.com/sharp378/paper-mc-ecs
-api-version: '1.20'
