+++ servinator-plugin/settings.gradle.kts
@@ -11,4 +11,4 @@ plugins {
 }
 
 rootProject.name = "servinator"
-include("app")
+include("ServinatorPlugin")
